Split up the filters file; changed a function to inline
